Revise the test placement based on review comments
